Restore TestMain in gateway_test package
