+++ pkg/gateway/main_test.go
@@ -1,0 +1,28 @@
+package gateway_test
+
+import (
+	"os"
+	"testing"
+
+	"github.com/ory/dockertest/v3"
+	"github.com/treeverse/lakefs/pkg/logging"
+	"github.com/treeverse/lakefs/pkg/testutil"
+)
+
+var (
+	pool        *dockertest.Pool
+	databaseURI string
+)
+
+func TestMain(m *testing.M) {
+	var err error
+	var closer func()
+	pool, err = dockertest.NewPool("")
+	if err != nil {
+		logging.Default().Fatalf("Could not connect to Docker: %s", err)
+	}
+	databaseURI, closer = testutil.GetDBInstance(pool)
+	code := m.Run()
+	closer() // cleanup
+	os.Exit(code)
+}
